requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
Async-OAuthlib==0.0.9
attrs==24.2.0
beautifulsoup4==4.12.3
biggo-api==0.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.20
langchain-openai==0.2.9
langgraph==0.2.53
langgraph-checkpoint==2.0.5
langgraph-sdk==0.1.36
langsmith==0.1.145
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
numpy==2.1.3
oauthlib==3.2.2
openai==1.55.0
openpyxl==3.1.5
orjson==3.10.11
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
propcache==0.2.0
pydantic==2.10.1
pydantic_core==2.27.1
pymongo==4.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xlrd==2.0.1
yarl==1.18.0
